Move element matrix code to separate file
