+++ PoroElastic/PoroElasticity.C
@@ -22,187 +22,6 @@
 #include "tinyxml.h"
 
 
-//! \brief Enum for element level solution vectors
-enum SolutionVectors
-{
-  U = 0,                        // Displacement
-  P = 1,                        // Pore pressure
-  NSOL = 2
-};
-
-
-//! \brief Enum for element level right-hand-side vectors
-enum ResidualVectors
-{
-  // System vectors, "global" size
-  Fsys = 0,                     // Final RHS vector
-  Fcur = 1,                     // RHS contribution from this timestep
-  Fprev = 2,                    // RHS contribution from previous timestep
-
-  // Sub-vectors, sized according to the bases in question
-  Fu = 3,                       // Traction and body forces
-  Fp = 4,                       // Flux and body forces
-
-  Fres = 5,
-  NVEC = 6
-};
-
-
-//! \brief Enum for element level left-hand-side matrices
-enum TangentMatrices
-{
-  // System matrices, "global" size (note: order is fixed according to Newmark API)
-  sys = 0,                      // Final Newton matrix
-  sys_mass = 1,                 // Mass (acceleration term in Newmark)
-  sys_gstiff = 2,               // Geometric stiffness (velocity term in Newmark)
-  sys_mstiff = 3,               // Material stiffness (zero-order term in Newmark)
-
-  // Sub-matrices, sized according to the bases in question
-  uu_K = 4,                     // Stiffness matrix
-  uu_M = 5,                     // Mass matrix
-  up = 6,                       // Coupling matrix
-  pp_S = 7,                     // Compressibility matrix
-  pp_P = 8,                     // Permeability matrix
-
-  NMAT = 9
-};
-
-
-PoroElasticity::Mats::Mats(size_t ndof_displ, size_t ndof_press, bool neumann)
-{
-  this->resize(NMAT, NVEC);
-  this->ndof_displ = ndof_displ;
-  this->ndof_press = ndof_press;
-
-  size_t ndof_tot = ndof_displ + ndof_press;
-
-  rhsOnly = neumann;
-  withLHS = !neumann;
-  b[Fsys].resize(ndof_tot);
-  b[Fprev].resize(ndof_tot);
-  b[Fu].resize(ndof_displ);
-  b[Fp].resize(ndof_press);
-  b[Fres].resize(ndof_tot);
-
-  if (!neumann) {
-    A[sys].resize(ndof_tot, ndof_tot);
-    A[sys_mass].resize(ndof_tot, ndof_tot);
-    A[sys_gstiff].resize(ndof_tot, ndof_tot);
-    A[sys_mstiff].resize(ndof_tot, ndof_tot);
-    A[uu_K].resize(ndof_displ, ndof_displ);
-    A[uu_M].resize(ndof_displ, ndof_displ);
-    A[up].resize(ndof_displ, ndof_press);
-    A[pp_S].resize(ndof_press, ndof_press);
-    A[pp_P].resize(ndof_press, ndof_press);
-  }
-}
-
-
-const Matrix& PoroElasticity::Mats::getNewtonMatrix() const
-{
-  return A[sys];
-}
-
-
-const Vector& PoroElasticity::Mats::getRHSVector() const
-{
-  return b[Fsys];
-}
-
-
-void PoroElasticity::MixedElmMats::add_uu(size_t source, size_t target, double scale)
-{
-  A[target].addBlock(A[source], scale, 1, 1);
-}
-
-
-void PoroElasticity::MixedElmMats::add_up(size_t source, size_t target, double scale)
-{
-  A[target].addBlock(A[source], scale, 1, 1 + ndof_displ);
-}
-
-
-void PoroElasticity::MixedElmMats::add_pu(size_t source, size_t target, double scale)
-{
-  A[target].addBlock(A[source], scale, 1 + ndof_displ, 1, true);
-}
-
-
-void PoroElasticity::MixedElmMats::add_pp(size_t source, size_t target, double scale)
-{
-  A[target].addBlock(A[source], scale, 1 + ndof_displ, 1 + ndof_displ);
-}
-
-
-void PoroElasticity::MixedElmMats::form_vector(const Vector &u, const Vector &p, size_t target)
-{
-  b[target] = u;
-  b[target].insert(b[target].end(), p.begin(), p.end());
-}
-
-
-void PoroElasticity::NonMixedElmMats::add_uu(size_t source, size_t target, double scale)
-{
-  Matrix& T = A[target];
-  Matrix& S = A[source];
-  size_t nsd = ndof_displ / ndof_press;
-  size_t nf = nsd + 1;
-
-  for (size_t i = 1; i <= ndof_press; ++i)
-    for (size_t j = 1; j <= ndof_press; ++j)
-      for (size_t l = 1; l <= nsd; ++l)
-        for (size_t k = 1; k <= nsd; ++k)
-          T(nf*(i-1)+l, nf*(j-1)+k) += scale * S(nsd*(i-1)+l, nsd*(j-1)+k);
-}
-
-
-void PoroElasticity::NonMixedElmMats::add_up(size_t source, size_t target, double scale)
-{
-  Matrix& T = A[target];
-  Matrix& S = A[source];
-  size_t nsd = ndof_displ / ndof_press;
-  size_t nf = nsd + 1;
-
-  for (size_t i = 1; i <= ndof_press; ++i)
-    for (size_t j = 1; j <= ndof_press; ++j)
-      for (size_t l = 1; l <= nsd; ++l)
-          T(nf*(i-1)+l, j*nf) += scale * S(nsd*(i-1)+l, j);
-}
-
-
-void PoroElasticity::NonMixedElmMats::add_pu(size_t source, size_t target, double scale)
-{
-  Matrix& T = A[target];
-  Matrix& S = A[source];
-  size_t nsd = ndof_displ / ndof_press;
-  size_t nf = nsd + 1;
-
-  for (size_t i = 1; i <= ndof_press; ++i)
-    for (size_t j = 1; j <= ndof_press; ++j)
-      for (size_t l = 1; l <= nsd; ++l)
-        T(j*nf, nf*(i-1)+l) += scale * S(nsd*(i-1)+l, j);
-}
-
-
-void PoroElasticity::NonMixedElmMats::add_pp(size_t source, size_t target, double scale)
-{
-  Matrix& T = A[target];
-  Matrix& S = A[source];
-  size_t nsd = ndof_displ / ndof_press;
-  size_t nf = nsd + 1;
-
-  for (size_t i = 1; i <= ndof_press; ++i)
-    for (size_t j = 1; j <= ndof_press; ++j)
-      T(i*nf, j*nf) += scale * S(i, j);
-}
-
-
-void PoroElasticity::NonMixedElmMats::form_vector(const Vector& u, const Vector& p, size_t target)
-{
-  utl::interleave(u, p, b[target], ndof_displ / ndof_press, 1);
-}
-
-
 PoroElasticity::PoroElasticity (unsigned short int n, int order) : Elasticity(n)
 {
   eS = Fu + 1;               // Elasticity::evalBou stores traction here
@@ -466,12 +285,12 @@ bool PoroElasticity::finalizeElement (LocalIntegral& elmInt, const TimeDomain& t
   elMat.add_pp(pp_P, sys, time.dt);
 
   // Construct the geometric stiffness matrix
-  elMat.add_pu(up, sys_gstiff);
-  elMat.add_pp(pp_S, sys_gstiff);
+  elMat.add_pu(up, sys_C);
+  elMat.add_pp(pp_S, sys_C);
 
   // Contribution to RHS from previous timestep
   elMat.form_vector(elmInt.vec[U], elmInt.vec[P], Fprev);
-  elMat.b[Fprev] = elMat.A[sys_gstiff] * elMat.b[Fprev];
+  elMat.b[Fprev] = elMat.A[sys_C] * elMat.b[Fprev];
 
   // Contribution to RHS from current timestep
   elMat.b[Fp] *= time.dt;
